MoonLive: process CodeRabbit review + raise script size limit
Processes the CodeRabbit review of PR #30 (Stage 1 controls) — fixing three real bugs plus a batch of hardening and doc findings — and raises the script source limit from 127 to 512 chars so a multi-control script (a declaration per control plus the statement) is not truncated. Verified live on both device ISAs: Xtensa (Olimex) and RISC-V (P4 — the first on-hardware run of the RISC-V control codegen, including the control value surviving a host call).

KPI: 16384lights | tick:130/88/120/9/1/320/39/16/20/64/132/11/34us(FPS:7692/11363/8333/111111/1000000/3125/25641/62500/50000/15625/7575/90909/29411) | ESP32:1270KB | tick:9105us(FPS:109) | heap:33216KB | src:118(22755) | test:76(12022) | lizard:84w

Core:
- MoonLiveCompiler: a control default must lie within its @control range (a slider can't start outside its bounds); a control name that shadows a built-in (e.g. random16) is rejected so name resolution can't collide with a call; names at/over kMaxControlName are rejected rather than silently truncated downstream.
- MoonLive: the control-values arena is allocated ONCE at full kMaxCtrls capacity and never reallocated, so a bound control pointer is fixed for the engine's lifetime (the previous grow-and-move could dangle a pointer published by a prior onBuildControls). The control set is published only after place() succeeds (a failed compile no longer advertises controls for code that isn't running), and a kept slot whose @control range shrank is clamped into the new bounds.
- MoonLiveIr: DeclaredControl uses uint8 min/max/def (validated 0..255) — 24 -> 16 bytes; kMaxControlName added.
- JsonUtil: parseString now also decodes \u00XX (the writer emits it for control bytes < 0x20), completing reader/writer escape symmetry alongside the named escapes.
- Control: addText/addTextArea take a uint16_t bufSize (was uint8_t, capping a text buffer at 255), and the apply path's maxLen + validator scratch widen to match — so a large multi-line value (a script source) round-trips intact.

Light domain:
- MoonLiveEffect: source buffer 128 -> 512 (a multi-control script fits; grow-on-demand for bigger Ripples-class scripts is backlogged). Control names are copied without truncation (the compiler guarantees they fit).

Core / platform:
- moonlive_asm_host / _riscv: register-map comments rewritten present-tense (no "became"/"left the pool"); the host Reg contract documents R4 as the controls pointer (kArg4).

Tests:
- unit_moonlive_compiler: malformed-decl matrix gains the default-outside-range and shadows-a-builtin cases.
- unit_JsonUtil_parse: \r \b \f and \u00XX decode assertions.
- scenario_MoonLiveEffect_controls: a two-controls -> one-control shrink step (orphan drop, remaining control stays bound).

Docs / CI:
- MoonLiveEffect.md: wire contract reworded — control names/ranges are derived from source and surfaced in /api/state (the device JSON view), not "missing" from it.
- decisions.md: the ok()-predicate (host test-path) and the parseString-symmetry / register-encoding (device-ISA) lessons split so they read as distinct fixes.
- livescripts-analysis-top-down.md: backlog the grow-on-demand source buffer; "Shipped so far" prose trimmed to the next-step + decisions link (a backlog shouldn't read like release history).
- Plan-20260627 - MoonLive Stage 1 controls.md renamed with the (shipped) outcome suffix.
- settings.local(.cleaned).json: broad allow patterns for the routine build/test/curl/edit loop (commits + destructive/external-network commands stay explicit).

Reviews:
- 🐇 control default outside @control range: fixed (rejected at decl).
- 🐇 control name shadows a builtin / ident-vs-call collision: fixed (rejected at decl).
- 🐇 control arena reallocates and dangles a bound pointer: fixed (allocate once at full capacity, never move).
- 🐇 control set published before place() succeeds: fixed (publish after).
- 🐇 a kept control's range change leaves a stale out-of-range value: fixed (clamp on recompile).
- 🐇 parseString not symmetric with the writer for \u00XX: fixed (decode it).
- 🐇 DeclaredControl wider than needed (int32 min/max/def): fixed (uint8, 24->16 bytes).
- 🐇 control name silently truncated into a prefix collision: fixed (reject overlong at compile).
- 🐇 asm register-map comments narrate changes / stale host Reg contract: fixed (present-tense; R4 = controls pointer).
- 🐇 backlog reads like release history / spec implies controls missing from JSON / decisions conflates two fixes / plan filename open-ended: fixed.
- 🐇 scenario missing a control-shrink transition: added.
- 🐇 scenario re-add measures a bare add, not a control source: accepted/deferred — construct-mode can't apply a dynamically-added scripted control as the final asserted render; control re-acquisition is already proven by the add-control-script step.
- 🐇 settings.local.cleaned.json over-broad allowlist + reverse-copy guard: accepted — product-owner-owned local tooling config (not project code); the cleaned<-private copy is the documented gate-8 restore step, and the convenience patterns re-accumulate per session.

+++ src/core/Control.h
@@ -243,7 +243,9 @@ public:
 
     // validate (optional): a per-control input check applied on every write path
     // (see ControlDescriptor::validate). nullptr accepts anything that fits the buffer.
-    void addText(const char* name, char* var, uint8_t bufSize = 16,
+    // bufSize is uint16_t so a large multi-line value (a script source, hundreds of bytes) isn't
+    // capped at 255; the descriptor's `max` (int32_t) carries it through the parse path.
+    void addText(const char* name, char* var, uint16_t bufSize = 16,
                  bool (*validate)(const char*) = nullptr) {
         grow();
         controls_[count_++] = {var, name, 0, ControlType::Text, 0, bufSize, false, false, validate};
@@ -251,7 +253,7 @@ public:
 
     // Like addText but the UI renders a resizable multi-line <textarea> (e.g. a
     // script source). Same char-buffer storage and parse/persist behaviour as Text.
-    void addTextArea(const char* name, char* var, uint8_t bufSize = 16,
+    void addTextArea(const char* name, char* var, uint16_t bufSize = 16,
                      bool (*validate)(const char*) = nullptr) {
         grow();
         controls_[count_++] = {var, name, 0, ControlType::TextArea, 0, bufSize, false, false, validate};
